media-source/SourceBuffer-appendBuffer-updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SouceBuffer#updating attribute after calling SourceBuffer#appendBuffer(data)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            var updatings = [];
            sourceBuffer.addEventListener('updatestart', function(e) {
                updatings.push(sourceBuffer.updating);
            });
            sourceBuffer.addEventListener('update', function(e) {
                updatings.push(sourceBuffer.updating);
            });
            sourceBuffer.addEventListener('updateend', function(e) {
                updatings.push(sourceBuffer.updating);
                t.step(function() {
                    assert_array_equals(updatings, 
                                        [false, true, true, false, false],
                                        'Check the transition of the updating attribute');
                });
                t.done();
            });
            updatings.push(sourceBuffer.updating);
            sourceBuffer.appendBuffer(data);
            updatings.push(sourceBuffer.updating);
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
}, 'SourceBuffer#appendBuffer(data) : Check the transition of updating attribute.');

</script>
</body>
</html>
